Crontab
Adding a rudimentary crontab. To apply this, first check that the cron
job isn't already scheduled. `crontab -l` will show the cron jobs for
the user. We can edit the user's cron jobs with `crontab -e`. Here you
should add these lines. The first is a comment and the second is an
instruction. Understanding the crontab file can be leared from the file
itself, it has a detailed explanation of each field.